+++ source/common/stream_info/stream_info_impl.h
@@ -4,6 +4,7 @@
 #include <cstdint>
 
 #include "envoy/common/time.h"
+#include "envoy/http/header_map.h"
 #include "envoy/stream_info/stream_info.h"
 
 #include "common/common/assert.h"
@@ -218,6 +219,25 @@ struct StreamInfoImpl : public StreamInfo {
     return upstream_transport_failure_reason_;
   }
 
+  void setRequestHeaders(const Http::HeaderMap& headers) override {
+    headers.iterate(
+        [](const Http::HeaderEntry& entry, void* self) {
+          static_cast<StreamInfoImpl*>(self)->request_headers_.emplace_back(
+              std::string(entry.key().getStringView()), entry.value().getStringView());
+          return Http::HeaderMap::Iterate::Continue;
+        },
+        this);
+  }
+
+  std::string getRequestHeader(const Http::LowerCaseString& name) const override {
+    for (const auto& entry : request_headers_) {
+      if (entry.first == name) {
+        return entry.second;
+      }
+    }
+    return std::string();
+  }
+
   void dumpState(std::ostream& os, int indent_level = 0) const {
     const char* spaces = spacesForLevel(indent_level);
     os << spaces << "StreamInfoImpl " << this << DUMP_OPTIONAL_MEMBER(protocol_)
@@ -255,6 +275,7 @@ private:
   Ssl::ConnectionInfoConstSharedPtr downstream_ssl_info_;
   Ssl::ConnectionInfoConstSharedPtr upstream_ssl_info_;
   std::string requested_server_name_;
+  Http::HeaderVector request_headers_;
   UpstreamTiming upstream_timing_;
   std::string upstream_transport_failure_reason_;
 };